==> notebooksBRIS/01_calibrate_cfsv2/05_convert_upslope_area2netcdf.ipynb <==
import os
import glob
import numpy as np
import pandas as pd
import xarray as xr
import rioxarray as rxr
from datetime import datetime, date
import json
from xgrads import open_CtlDataset
import geopandas as gpd
import requests
import matplotlib.pyplot as plt
from geocube.api.core import make_geocube
%matplotlib inline

# In QGIS

Once you have the huc4 geojson, the streamgage station shapefiles and the filled dem downloaded - you need to do some work in QGIS to determine the upslope area of each stream gage. 

A detailed description of the workflow can be found here: https://docs.google.com/document/d/1Urdpp41tCZp0ZptHDdRD2HQ3ZznolhWA0Zaa_szcgfc/edit?usp=sharing
Here is the overview:
1. Load DEM, huc4 shapefile, and gage shapefile
2. Reproject huc4 shape to match DEM
2. Clip dem to HUC4
3. Fill Sinks tool
4. Strahler Order tool
5. Using station locations, use Upslope Area tool to save out .tif files with upslope area for each gage. 
6. Rasterize the huc4 shapefile to use for masking later. 

# Choose a domain
domain = 'BRIS'

# Path to the SnowModel folder
gagepath = '/nfs/attic/dfh/2020_NPRB/data/stream_gage/'+domain+'/'

#path to NPRB domains
domains_resp = requests.get("https://raw.githubusercontent.com/NPRB/02_preprocess_python/main/NPRB_domains.json")
domains = domains_resp.json()

list_of_files = sorted( filter( os.path.isfile,
                        glob.glob(gagepath + '*.tif') ) )

list_of_files

for i in range(len(list_of_files)):

    file = list_of_files[i]
    stncode=file[-12:-4]
    data = rxr.open_rasterio(file)
    da = data.sel(band=1)
    if i ==0:
        ds = da.to_dataset(name=stncode)
    else:
        ds[stncode]=da
ds

ds.to_netcdf(gagepath+'upslope.nc')
